requirements.txt
# Ruya Hackathon — core stack (ClickHouse for storage)
python-dotenv>=1.0.0
streamlit>=1.28.0
google-generativeai>=0.3.0
groq>=0.4.0
openai>=1.0.0
elevenlabs>=1.0.0
clickhouse-connect>=0.7.0
langfuse>=2.0.0
